Add avalon_commit var to set git hash for build
Create an optional avalon_commit TF var to allow manually
setting which git commit to base the CodeBuild off of. If
unset (the default), will operate as previously, pulling
the commit hash from the branch name.

Includes a rename of internal var from AVALON_REV to
AVALON_COMMIT to consistency between it and the var.