Use correct device for Solr data and fail safely.
This was using a nonexistent device for the Solr data. An alias to the
correct device is declared as the 'device_name' attribute on
'aws_ebs_volume.solr_data', so this uses that instead. This also makes
the mountpoint unwriteable if it can't be mounted, so Solr will fail
rather than writing its data to the EC2 instance's root filesystem.