compose: Rebuild the EC2 VM as needed, etc.
The VM should be rebuilt when its configuration changes or when there's
a newer VM image. Also, since the "user data" script isn't idempotent,
the VM should be rebuilt rather than restarted when the script changes.
Lowering the TTL on the VM's A record, which is only used for
administrative access, will make it available faster without imposing a
burden on DNS.